+++ src/routing/mod.rs
@@ -1,0 +1,7 @@
+//! Actor routing graph and commitment-planning algorithms.
+
+mod commitment;
+mod graph;
+
+pub use commitment::*;
+pub use graph::*;
